+++ examples/sample3/scene/background.go
@@ -48,9 +48,9 @@ func (bg *Background) getSpeed() float64 {
 func (bg *Background) move() {
 	p := bg.GetPosition()
 	s := bg.GetScale()
-	p.X -= (int)(bg.speed)
+	p.X -= float32(bg.speed)
 	bg.SetPositionX(p.X)
 	if p.X < -1*s.W/2 {
-		bg.SetPositionX((int)(config.ScreenWidth/2 + (config.ScreenWidth - float32(bg.speed))))
+		bg.SetPositionX(config.ScreenWidth/2 + (config.ScreenWidth - float32(bg.speed)))
 	}
 }
